Ignore all target directories.
When working in an IDE like vscode, it uses rust infrastructure to
check the code, which results in "target" folders in the base of
the crates I am working on. For example, if I am working on pcid
code, I get a "target" folder in the pcid folder. This change
causes git to ignore all those target directories.

Signed-off-by: Wren Turkal <wt@penguintechs.org>